Naver Checkstyle 포장지에 적용 (#346)
