+++ src/main/java/shop/nuribooks/books/order/wrapping/dto/WrappingPaperRequest.java
@@ -1,26 +1,26 @@
 package shop.nuribooks.books.order.wrapping.dto;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import shop.nuribooks.books.order.wrapping.entity.WrappingPaper;
 
-import java.math.BigDecimal;
-
 public record WrappingPaperRequest(
-        @NotBlank(message = "포장지 이름은 필수입니다.")
-        String title,
-        @NotBlank(message = "포장지 이미지 경로는 필수입니다.")
-        String imageUrl,
-        @NotNull(message = "포장지 가격은 필수입니다.")
-        @Min(value = 0, message = "포장지 가격은 0원 이상이어야합니다.")
-        BigDecimal wrappingPrice
+	@NotBlank(message = "포장지 이름은 필수입니다.")
+	String title,
+	@NotBlank(message = "포장지 이미지 경로는 필수입니다.")
+	String imageUrl,
+	@NotNull(message = "포장지 가격은 필수입니다.")
+	@Min(value = 0, message = "포장지 가격은 0원 이상이어야합니다.")
+	BigDecimal wrappingPrice
 ) {
-    public WrappingPaper toEntity() {
-        return WrappingPaper.builder()
-                .title(title)
-                .imageUrl(imageUrl)
-                .wrappingPrice(wrappingPrice)
-                .build();
-    }
+	public WrappingPaper toEntity() {
+		return WrappingPaper.builder()
+			.title(title)
+			.imageUrl(imageUrl)
+			.wrappingPrice(wrappingPrice)
+			.build();
+	}
 }
