+++ src/main/java/shop/nuribooks/books/order/wrapping/dto/WrappingRegisterRequest.java
@@ -1,5 +1,4 @@
 package shop.nuribooks.books.order.wrapping.dto;
 
-// todo : 포장 요청 정보
 public record WrappingRegisterRequest(long wrappingPaperId) {
 }
