+++ src/main/java/shop/nuribooks/books/order/wrapping/repository/WrappingPaperRepository.java
@@ -1,6 +1,7 @@
 package shop.nuribooks.books.order.wrapping.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
 import shop.nuribooks.books.order.wrapping.entity.WrappingPaper;
 
 public interface WrappingPaperRepository extends JpaRepository<WrappingPaper, Long> {
