+++ src/main/java/shop/nuribooks/books/order/wrapping/service/WrappingService.java
@@ -1,4 +1,4 @@
-package shop.nuribooks.books.order.wrapping.Service;
+package shop.nuribooks.books.order.wrapping.service;
 
 /**
  * 주문에 대한 포장 정보 서비스 인터페이스
